Move RangeSliderThumb into its own file
